renamed files based on other's commands feedback
